requirements.txt
streamlit
pandas
scikit-learn
pyserial
firebase-admin
